+++ CHANGES.rst
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+0.6.0 (unreleased)
+------------------
+
+* `#217 <https://github.com/jantman/awslimitchecker/issues/217>`_ - add support
+  for new/missing EC2 instance types: ``m4.16xlarge``, ``x1.16xlarge``, ``x1.32xlarge``,
+  ``p2.xlarge``, ``p2.8xlarge``, ``p2.16xlarge``.
+* `#215 <https://github.com/jantman/awslimitchecker/issues/215>`_ - support
+  "Regional Benefit" Reserved Instances that have no specific AZ set on them (
+  bug fix for a KeyError when calculating EC2 Reserved Instances). It's unclear
+  to me whether, for the purpose of calculating limits, these are subtracted from
+  On-Demand Running Instances or ignored (because they're not actually a specific
+  capacity reservation). For the time being, they'll be ignored (i.e. not subtracted
+  from On-Demand Running Instances) but a debug-level message will be logged
+  for each.
+
 0.5.1 (2016-09-25)
 ------------------